Promisify authorize() method, cleanup
